refactor: rename adapter and middleware
