+++ packages/dubbo-express/src/index.ts
@@ -12,4 +12,4 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-export { expressConnectMiddleware } from "./express-connect-middleware.js";
+export { expressConnectMiddleware } from "./express-dubbo-middleware.js";
