+++ packages/dubbo-fastify/src/index.ts
@@ -12,4 +12,4 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-export { fastifyConnectPlugin } from "./fastify-connect-plugin.js";
+export { fastifyConnectPlugin } from "./fastify-dubbo-plugin.js";
