+++ packages/dubbo-next/src/index.ts
@@ -12,4 +12,4 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-export { nextJsApiRouter } from "./connect-nextjs-adapter.js";
+export { nextJsApiRouter } from "./dubbo-nextjs-adapter.js";
