+++ packages/dubbo-node/src/index.ts
@@ -19,7 +19,7 @@ export { createGrpcWebTransport } from "./grpc-web-transport.js";
 export { createGrpcTransport } from "./grpc-transport.js";
 export { createConnectTransport } from "./dubbo-transport.js";
 export { compressionBrotli, compressionGzip } from "./compression.js";
-export { connectNodeAdapter } from "./connect-node-adapter.js";
+export { connectNodeAdapter } from "./dubbo-node-adapter.js";
 
 export {
   universalRequestFromNodeRequest,
